feature: add LAN_X command encoders for loco control, turnout management, and track power
